Provide own version comparator
LibreOffice's python does not come with setuptools or packaging by default, and not even pip, so replacing distutils with packaging or using setuptools, while doable, poses quite some setup to get this running, for little merit.

Since I cannot find any evidence of letters being used in LO's versioning schemes (old or new), I think/hope this comparator should do the trick just fine.

NOTE that it is being used in a test to check whether we're not using 'LibreOffice' as such, or whether the version is as old as 3.3 or below. Pretty old.